Files: relocated About.js into view folder. #2
